Fix Button onCheckedClick
Ran into an issue for over the past day where none of the binding attributes were making it to the views within the home fragment.  Finally something changed and they are receiving the values and binds.  Not sure what it was, maybe the context within the XML, maybe it was adding things to the binding like the viewModel and viewLifecycleOwner even though those did not immediately fix the issue.

Either way, the buttons kind of work and I would just like to move on from here.