Interpret and display numbers
Will identify a number key and represent the following letters as numbers.  Also reverts back to letters if encountering letters past J, does not yet display number related punctuation like commas and decimals.
Had to convert the for loop to an indexed while loop so that the mode can be changed and then the loop run again on the same Letter.