Sound Playback on button toggles
Uses short audio files and the android.media.SoundPool to trigger audio feedback for button toggles.  Opted to use the deprecated SoundPool since it still fit this non-essential use case better than the recommended MediaPlayer option which would require creation and deletion of mp each time we want to play a sound. This just seemed like the better option. Still need to add a mute option for the sounds next.