Move Translation to Braille.kt and add Final Abbrev
Moved the LiveData transformation to its own function within the Braille companion object.  Also added more helper functions to change to a final abbreviation if detected at the end of a word.  This required moving the capitalization logic since the final abbreviation 3 is made up of a capital N or Y at the end of a word.  There might still be bugs if the abbreviation is not actually at the end of a word, but so far detection is okay for normal usage.