Change animations to build after layout is set
Had to extract the animation building to its own function and have that posted as a runnable to the main layout so the view positions it requires are properly laid out. Trying to determine these in any of Android's many onCreated functions does not work since the views are created but not finalized until an indeterminate time later.

Signed-off-by: atakemori <atakemori99@gmail.com>