Add Privacy Policy and dialog
Added a privacy policy to the git repo and as an xml string resource within the app.  While the other preferences currently link to github links for the license and feedback/issues, the privacy policy is popped up in-app as an alertDialog.  Later might want to differentiate between preference onTouch for alertDialogs, and onLongTouch for launching intents out-of-app.

Also removed Dark theme entry for now, so only Light is available.

Signed-off-by: atakemori <atakemori99@gmail.com>